Don't use pnstrdup() to duplicate non-string data area.
Previously pnstrdup() was used to duplicate the cached result of pgs2norm().
This was improper usage because that cached result is not string. This and
the change of pnstrdup() by the commit 82c117cb90 in PostgreSQL caused
pgs2norm() to emit a segmentation fault.

This commit changes pgs2norm() so that it duplicates the cached result by
using the original logic instead of pnstrdup().