Make regression test for textporter work fine on PostgreSQL 12 or later.
Previously the regression test for textporter failed when it ran on
PostgreSQL 12 or later. The cause of this issue is that commit d8594d1
in PostgreSQL changed the outputs of textporter regression test results.
This leads the existing expected file that depends on PostgreSQL 11
or before not to match with the output of regression test on 12,
and then caused the regression test to fail.

To fix the above issue, this commit adds textporter_1.source as
a variant comparison file, which contains the expected output of
textporter regression test on 12 or later.

Author: Torikoshi Atsushi
Reviewed-by: Fujii Masao
Discussion: https://osdn.net/projects/ludiafuncs/lists/archive/hackers/2020-February/000040.html