Support terse logging in enable_debug.
This commit allows us to set enable_debug option to 'terse' as
a setting value. When it's 'terse', the contents dealt in ludia_funcs
internal functions are excluded from the debug messages.
For example, if it's set to 'on', not only the name of function
pgs2norm() but also the target string that the function normalizes
are included in the debug messages. But if it's 'terse', only the
terse information like function name is output.

Previously we set enable_debug to 'on', in order to check
whether recheck is skipped or not in the regression test.
But the order of the contents in the debug messages varied
depending on the block size in PostgreSQL server. Which
would lead to unexpected failure of regression test in PostgreSQL
built with non-default block size. In order to stabilize the result
of this regression test, this patch make the test use debug level
'terse'.

Author: Torikoshi Atsushi
Reviewed-by: Fujii Masao
Discussion: https://osdn.net/projects/ludiafuncs/lists/archive/hackers/2020-February/000046.html